Fix song title newline
The title was being read from yt-dlp including a newline, since the title was at the end of what was written to the stdout pipe. Since that newline was not trimmed, it was included as part of the title and had effects in the web interface and queue. A <br> element would be at the end of each song title, but luckily that didn't have a visual effect. When it came to queueing songs, song info only ended up on individual lines because the title and its newline was at the end of each line. After fixing the newline problem, an additional newline had to be written every time queued song data was written to the queue file.