Add events system for pngbinds
An "event" being a set of media that can be displayed instead of the usual pngbinds state, given that it's conditions are met and that the state allows for events to interrupt it (by default this is not allowed). I'll stil need to add to the interface so that events can be easily managed.